+++ NEWS.md
@@ -2,7 +2,10 @@
 
 ## Version 2.3-1, July 11, 2022
 
-* 
+* The strict uniqueness of n.clones in dc.fit can be relaxed 
+  by the new check.nclones argument.
+* The update argument of dc.fit can be of length 1 or more,
+  updatefun must return a list when length(update) > 1.
 
 ## Version 2.3-0, Mar 21, 2019